refactor(admin): split AdminDashboardClient into sub-components. Fixes #369
Extracted the 646-line monolith into focused sub-components:
- admin-types.ts — shared types (AdminUser, SortField, SortDir) and helpers
- AdminSortableHeader.tsx — reusable sortable column header with SortIcon
- AdminSearchBar.tsx — search/filter input with result count
- AdminStatsCards.tsx — tier summary cards grid
- AdminUserTable.tsx — user table with rows, imports AdminSortableHeader

AdminDashboardClient.tsx is now 219 lines and acts as the orchestrator:
state management, data fetching, event listeners, and layout composition.

Pure refactor — no functional changes. All existing tests pass with minimal
updates (source-based regex tests now read from ALL admin source files).
